requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.2
fpdf==1.7.2
pyTelegramBotAPI==4.26.0